feat: Add Deno Update GitHub Action
This commit introduces a new GitHub Action workflow named 'Deno Update'.
It is scheduled to run daily at 9 AM UTC and can also be manually
triggered. The workflow runs on the latest Ubuntu and has write-all
permissions. It uses actions/checkout@v4 to checkout the code and
hasundue/molt-action@v1-rc to perform the update.